requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
httpx
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
itsdangerous
starlette
